feat : 외부호출 및 서킷브레이커 적용, 주문과 결제 트랜잭션 분리

+++ apps/commerce-api/src/main/kotlin/com/loopers/domain/product/ProductEntity.kt
@@ -10,7 +10,6 @@ import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
 import org.hibernate.annotations.SQLDelete
 import org.hibernate.annotations.SQLRestriction
-import java.math.BigDecimal
 
 @SQLRestriction("deleted_at is null")
 @SQLDelete(sql = "update product set deleted_at = CURRENT_TIMESTAMP where id = ?")
@@ -19,7 +18,7 @@ import java.math.BigDecimal
 class ProductEntity(
     name: String,
     brandId: Long,
-    price: BigDecimal,
+    price: Long,
     stock: Int = 0
 ) : BaseEntity() {
 
@@ -34,7 +33,7 @@ class ProductEntity(
     var brand: BrandEntity? = null
 
     @Column(name = "price")
-    val price: BigDecimal = price
+    val price: Long = price
 
     @Column(name = "stock")
     var stock: Int = stock
