+++ apps/commerce-api/src/test/kotlin/com/loopers/fixture/order/OrderFixture.kt
@@ -2,7 +2,6 @@ package com.loopers.fixture.order
 
 import com.loopers.domain.type.PaymentType
 import com.loopers.interfaces.api.order.OrderV1Models
-import java.math.BigDecimal
 
 sealed class OrderFixture {
 
@@ -14,14 +13,14 @@ sealed class OrderFixture {
                 items = listOf(
                     OrderV1Models.Request.Item(
                         productId = 1L,
-                        price = BigDecimal.valueOf(1000),
+                        price = 1000L,
                         qty = 2
                     )
                 ),
                 payments = listOf(
                     OrderV1Models.Request.Payment(
                         type = PaymentType.POINT,
-                        amount = BigDecimal.valueOf(1000),
+                        amount = 1000L,
                     )
                 )
             )
@@ -34,19 +33,19 @@ sealed class OrderFixture {
                 items = listOf(
                     OrderV1Models.Request.Item(
                         productId = 1L,
-                        price = BigDecimal.valueOf(1000),
+                        price = 1000L,
                         qty = 1
                     ),
                     OrderV1Models.Request.Item(
                         productId = 2L,
-                        price = BigDecimal.valueOf(2000),
+                        price = 2000L,
                         qty = 3
                     )
                 ),
                 payments = listOf(
                     OrderV1Models.Request.Payment(
                         type = PaymentType.POINT,
-                        amount = BigDecimal.valueOf(1000),
+                        amount = 1000L,
                     )
                 )
             )
@@ -59,14 +58,14 @@ sealed class OrderFixture {
                 items = listOf(
                     OrderV1Models.Request.Item(
                         productId = 1L,
-                        price = BigDecimal.valueOf(-100),
+                        price = -100L,
                         qty = 1
                     )
                 ),
                 payments = listOf(
                     OrderV1Models.Request.Payment(
                         type = PaymentType.POINT,
-                        amount = BigDecimal.valueOf(1000),
+                        amount = 1000L,
                     )
                 )
             )
@@ -79,14 +78,14 @@ sealed class OrderFixture {
                 items = listOf(
                     OrderV1Models.Request.Item(
                         productId = 1L,
-                        price = BigDecimal.valueOf(100),
+                        price = 100L,
                         qty = -1
                     )
                 ),
                 payments = listOf(
                     OrderV1Models.Request.Payment(
                         type = PaymentType.POINT,
-                        amount = BigDecimal.valueOf(1000),
+                        amount = 1000L,
                     )
                 )
             )
@@ -100,7 +99,7 @@ sealed class OrderFixture {
                 payments = listOf(
                     OrderV1Models.Request.Payment(
                         type = PaymentType.POINT,
-                        amount = BigDecimal.valueOf(1000),
+                        amount = 1000L,
                     )
                 )
             )
